+++ src/main/java/dev/enjarai/trickster/spell/tricks/entity/query/GetFacingTrick.java
@@ -1,26 +1,28 @@
-package dev.enjarai.trickster.spell.tricks.entity;
+package dev.enjarai.trickster.spell.tricks.entity.query;
 
 import dev.enjarai.trickster.spell.Fragment;
 import dev.enjarai.trickster.spell.Pattern;
 import dev.enjarai.trickster.spell.SpellContext;
-import dev.enjarai.trickster.spell.fragment.BooleanFragment;
 import dev.enjarai.trickster.spell.fragment.FragmentType;
+import dev.enjarai.trickster.spell.fragment.VectorFragment;
 import dev.enjarai.trickster.spell.tricks.Trick;
 import dev.enjarai.trickster.spell.tricks.blunder.BlunderException;
 import dev.enjarai.trickster.spell.tricks.blunder.UnknownEntityBlunder;
+import org.joml.Vector3d;
 
 import java.util.List;
 
-public class SneakingReflectionTrick extends Trick {
-    public SneakingReflectionTrick() {
-        super(Pattern.of(2, 4, 7));
+public class GetFacingTrick extends Trick {
+    public GetFacingTrick() {
+        super(Pattern.of(3, 2, 7));
     }
 
     @Override
     public Fragment activate(SpellContext ctx, List<Fragment> fragments) throws BlunderException {
         var entity = expectInput(fragments, FragmentType.ENTITY, 0);
 
-        return new BooleanFragment(entity.getEntity(ctx)
-                .orElseThrow(() -> new UnknownEntityBlunder(this)).isSneaking());
+        var facing = entity.getEntity(ctx).orElseThrow(() -> new UnknownEntityBlunder(this)).getRotationVector();
+
+        return new VectorFragment(new Vector3d(facing.x, facing.y, facing.z));
     }
 }
