+++ src/main/java/dev/enjarai/trickster/spell/tricks/entity/query/RaycastBlockSideTrick.java
@@ -1,4 +1,4 @@
-package dev.enjarai.trickster.spell.tricks.entity;
+package dev.enjarai.trickster.spell.tricks.entity.query;
 
 import dev.enjarai.trickster.spell.Fragment;
 import dev.enjarai.trickster.spell.Pattern;
@@ -10,14 +10,14 @@ import net.minecraft.util.hit.HitResult;
 
 import java.util.List;
 
-public class RaycastBlockPosTrick extends AbstractRaycastBlockTrick {
-    public RaycastBlockPosTrick() {
-        super(Pattern.of(3, 4, 5, 2, 4));
+public class RaycastBlockSideTrick extends AbstractRaycastBlockTrick {
+    public RaycastBlockSideTrick() {
+        super(Pattern.of(3, 4, 5, 2, 4, 8, 5));
     }
 
     @Override
     public Fragment activate(SpellContext ctx, List<Fragment> fragments) throws BlunderException {
         var hit = getHit(ctx, fragments);
-        return hit.getType() == HitResult.Type.MISS ? VoidFragment.INSTANCE : VectorFragment.of(hit.getBlockPos());
+        return hit.getType() == HitResult.Type.MISS ? VoidFragment.INSTANCE : VectorFragment.of(hit.getSide().getVector());
     }
 }
